refactor(instrumentation): rename constructors and ifaces to idomatic Go
