+++ internal/instrumentation/exporter.go
@@ -1,46 +1,66 @@
 // Copyright 2020 New Relic Corporation. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
+
 package instrumentation
 
 import (
+	"context"
 	"net/http"
+
+	"github.com/prometheus/client_golang/prometheus"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+	"go.opentelemetry.io/otel/api/metric"
+	oprometheus "go.opentelemetry.io/otel/exporters/metric/prometheus"
+	"go.opentelemetry.io/otel/label"
 )
 
-type MetricType int
+type instrumentation struct {
+	handler  *oprometheus.Exporter
+	meter    *metric.Meter
+	counters map[MetricName]metric.Int64Counter
+}
 
-const (
-	Counter MetricType = iota
-	Gauge
-)
+func (i instrumentation) GetHandler() http.Handler {
+	return i.handler
+}
 
-type MetricName int
+func (i instrumentation) Measure(metricType MetricType, name MetricName, val int64) {
+	i.meter.RecordBatch(
+		context.Background(),
+		[]label.KeyValue{},
+		i.counters[name].Measurement(val))
+}
 
-const (
-	DMRequestsForwarded MetricName = iota // integration payload received
-	DMDatasetsReceived
-	EntityRegisterEntitiesRegistered
-	EntityRegisterEntitiesRegisteredWithWarning
-	EntityRegisterEntitiesRegistrationFailed
-	LoggedErrors
-)
+func (i instrumentation) GetHttpTransport(base http.RoundTripper) http.RoundTripper {
+	return otelhttp.NewTransport(base,
+		otelhttp.WithMeterProvider(i.handler.MeterProvider()),
+		otelhttp.WithMessageEvents(
+			otelhttp.ReadEvents,
+			otelhttp.WriteEvents))
+}
 
-var (
-	// NoopMeasure no-op Measure function type.
-	NoopMeasure       = func(_ MetricType, _ MetricName, _ int64) {}
-	metricsToRegister = map[MetricName]string{
-		DMRequestsForwarded:                         "dm.requests_forwarded",
-		DMDatasetsReceived:                          "dm.datasets_received",
-		EntityRegisterEntitiesRegistered:            "entity_register.entities_registered",
-		EntityRegisterEntitiesRegisteredWithWarning: "entity_register.entities_registered_with_warning",
-		EntityRegisterEntitiesRegistrationFailed:    "entity_register.entities_registration_failed",
-		LoggedErrors:                                "logged.errors",
+// New creates a new instrumentation bundle (exporter + measure fn...).
+func New() (Instrumenter, error) {
+	registry := prometheus.NewRegistry()
+	registry.MustRegister(prometheus.NewProcessCollector(prometheus.ProcessCollectorOpts{}))
+	registry.MustRegister(prometheus.NewGoCollector())
+	prometheusExporter, err := oprometheus.InstallNewPipeline(oprometheus.Config{
+		Registry: registry,
+	})
+	if err != nil {
+		return nil, err
 	}
-)
+	meter := prometheusExporter.MeterProvider().Meter("newrelic.infra")
 
-type Measure func(metricType MetricType, name MetricName, val int64)
+	counters := make(map[MetricName]metric.Int64Counter, 2)
 
-type Instrumentation interface {
-	GetHandler() http.Handler
-	Measure(metricType MetricType, name MetricName, val int64)
-	GetHttpTransport(base http.RoundTripper) http.RoundTripper
+	for metricName, metricRegistrationName := range metricsToRegister {
+		counters[metricName] = metric.Must(meter).NewInt64Counter("newrelic.infra/instrumentation." + metricRegistrationName)
+	}
+
+	return &instrumentation{
+		handler:  prometheusExporter,
+		counters: counters,
+		meter:    &meter,
+	}, err
 }
